Make transfer_stage.py be a wrapper script to a package module
